+++ .semversioner/next-release/patch-20240830181135475287.json
@@ -1,0 +1,4 @@
+{
+  "type": "patch",
+  "description": "Add entrypoints for incremental indexing"
+}
