+++ graphrag/index/api.py
@@ -8,8 +8,6 @@ WARNING: This API is under development and may undergo changes in future release
 Backwards compatibility is not guaranteed at this time.
 """
 
-from pathlib import Path
-
 from graphrag.config import CacheType, GraphRagConfig
 
 from .cache.noop_pipeline_cache import NoopPipelineCache
@@ -24,8 +22,10 @@ from .typing import PipelineRunResult
 
 async def build_index(
     config: GraphRagConfig,
-    run_id: str,
-    memory_profile: bool,
+    run_id: str = "",
+    is_resume_run: bool = False,
+    is_update_run: bool = False,
+    memory_profile: bool = False,
     progress_reporter: ProgressReporter | None = None,
     emit: list[str] | None = None,
 ) -> list[PipelineRunResult]:
@@ -37,6 +37,10 @@ async def build_index(
         The configuration.
     run_id : str
         The run id. Creates a output directory with this name.
+    is_resume_run : bool default=False
+        Whether to resume a previous index run.
+    is_update_run : bool default=False
+        Whether to update a previous index run.
     memory_profile : bool
         Whether to enable memory profiling.
     progress_reporter : ProgressReporter | None default=None
@@ -50,7 +54,10 @@ async def build_index(
     list[PipelineRunResult]
         The list of pipeline run results
     """
-    resume = Path(config.storage.base_dir).exists()
+    if is_resume_run and is_update_run:
+        msg = "Cannot resume and update a run at the same time."
+        raise ValueError(msg)
+
     pipeline_config = create_pipeline_config(config)
     pipeline_cache = (
         NoopPipelineCache() if config.cache.type == CacheType.none is None else None
@@ -63,7 +70,8 @@ async def build_index(
         cache=pipeline_cache,
         progress_reporter=progress_reporter,
         emit=([TableEmitterType(e) for e in emit] if emit is not None else None),
-        is_resume_run=resume,
+        is_resume_run=is_resume_run,
+        is_update_run=is_update_run,
     ):
         outputs.append(output)
         if progress_reporter:
